425project/requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
librosa>=0.10.0
pandas>=2.0.0
umap-learn>=0.5.3
scipy>=1.10.0
tqdm>=4.65.0
Pillow>=10.0.0
soundfile>=0.12.0
tabulate>=0.9.0